+++ examples/tap/tap_consumer.ex
@@ -1,16 +1,21 @@
 defmodule TapConsumer do
-  @behaviour Drinkup.Tap.Consumer
+  use Drinkup.Tap,
+    name: MyTap,
+    host: "http://localhost:2480"
 
+  @impl true
   def handle_event(%Drinkup.Tap.Event.Record{} = record) do
     IO.inspect(record, label: "Tap record event")
+    :ok
   end
 
   def handle_event(%Drinkup.Tap.Event.Identity{} = identity) do
     IO.inspect(identity, label: "Tap identity event")
+    :ok
   end
 end
 
-defmodule TapExampleSupervisor do
+defmodule ExampleTapConsumer do
   use Supervisor
 
   def start_link(arg \\ []) do
@@ -19,14 +24,7 @@ defmodule TapExampleSupervisor do
 
   @impl true
   def init(_) do
-    children = [
-      {Drinkup.Tap,
-       %{
-         consumer: TapConsumer,
-         name: MyTap,
-         host: "http://localhost:2480"
-       }}
-    ]
+    children = [TapConsumer]
 
     Supervisor.init(children, strategy: :one_for_one)
   end